Fix module names that are not valid identifiers (#1226)
This regressed in https://github.com/geldata/gel-js/pull/900 due to not
having a good test case. The way to reproduce this issue is to ensure
that the module ends up as something that the `schema` module would need
to import, and it seems like that requires object types defined in these
edge case modules.