Return identity_id, if available, from signup
In 6.0, we started returning the identity_id from signup even if
email verification was required. This allows developers to create User
objects connected to the newly created identity, even before the User
verifies their email. It also allows developers to treat "missing PKCE"
in the email verification flow as a success state that requires the
end user to simply sign in again. This was difficult to do before
because we treated the verification flow as the end of the "sign up"
flow, but if you want to allow no-PKCE verification, you need to
sign up the user before verfication is complete.