Optimize scalar cast maps (#1268)
Fixes #1242.

## `scalarCastableFrom`/`scalarAssignableBy`

This changes the `scalarCastableFrom`/`scalarAssignableBy` types to just
pick the type based on the given type's FQN.
This is much easier for TS, esp with generated enums, as their members
are statically written as well and TS was duck checking all of these.
```diff
- export type scalarAssignableBy<T extends $.ScalarType> =
- T extends _std.number ? _std.number :
- T extends _std.int64 ? _std.int64 :
- ...
+ export interface ScalarAssignableByMap {
+   "std::number": _std.$number;
+   "std::int64": _std.$int64;
+   ...
+ }
```

## `getSharedParentScalar`

This also changes `getSharedParentScalar` to compare against the types'
FQN.
This should be much more performant and still safe since these are
scalar types with unique names.
We can also skip enumerating every case if `A` & `B` are the same.
This saves having to declare mapping for 80% of the scalars that can't
be cast into another.

This removes 79 branches with my generated client, leaving only 6
specific scalar branches.
Before:
```ts
type getSharedParentScalar<A, B> =
  A extends _cal.$local_date ?
    B extends _cal.$local_datetime ?
      B
    :
    B extends _cal.$local_date ?
      B
    :
    never
  :
  A extends _std.$number ?
    B extends _std.$number ?
      B
    :
    never
  :
  A extends _sys.$VersionStage ?
    B extends _sys.$VersionStage ?
      B
    :
    never
  :
  ...
```
After:
```ts
type getSharedParentScalar<A, B> =
  AType extends $.ScalarType<infer AName> ?
  BType extends $.ScalarType<infer BName> ?
    AName extends "cal::local_date" ?
      BName extends "cal::local_datetime" ?
        B
      :
      BName extends "cal::local_date" ?
        B
      :
      never
    :
    ...
    AName extends BName ? A : never
```